Db work and types

+++ fpl_common/src/types/club_id.rs
@@ -4,21 +4,21 @@ use std::fmt::Display;
 use std::ops::Deref;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-#[serde(try_from = "i32")]
+#[serde(try_from = "i16")]
 #[derive(sqlx::Type)]
 #[sqlx(transparent)]
-pub struct ClubId(i32);
+pub struct ClubId(i16);
 
 #[derive(Debug, thiserror::Error)]
 #[error("ClubId must be between 1 and 20, got {0}")]
-pub struct ClubIdError(i32);
+pub struct ClubIdError(i16);
 
 impl ClubId {
-    pub const MIN: i32 = 1;
-    pub const MAX: i32 = 20;
-    pub const ALL: std::ops::RangeInclusive<i32> = Self::MIN..=Self::MAX;
+    pub const MIN: i16 = 1;
+    pub const MAX: i16 = 20;
+    pub const ALL: std::ops::RangeInclusive<i16> = Self::MIN..=Self::MAX;
 
-    pub fn new(club_id: i32) -> Result<Self, ClubIdError> {
+    pub fn new(club_id: i16) -> Result<Self, ClubIdError> {
         if Self::ALL.contains(&club_id) {
             Ok(Self(club_id))
         } else {
@@ -28,7 +28,7 @@ impl ClubId {
 }
 
 impl Deref for ClubId {
-    type Target = i32;
+    type Target = i16;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -41,27 +41,27 @@ impl Display for ClubId {
     }
 }
 
-impl TryFrom<i32> for ClubId {
+impl TryFrom<i16> for ClubId {
     type Error = ClubIdError;
 
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
+    fn try_from(value: i16) -> Result<Self, Self::Error> {
         Self::new(value)
     }
 }
 
-impl From<ClubId> for i32 {
+impl From<ClubId> for i16 {
     fn from(id: ClubId) -> Self {
         id.0
     }
 }
 
-impl PartialEq<i32> for ClubId {
-    fn eq(&self, other: &i32) -> bool {
+impl PartialEq<i16> for ClubId {
+    fn eq(&self, other: &i16) -> bool {
         self.0 == *other
     }
 }
 
-impl PartialEq<ClubId> for i32 {
+impl PartialEq<ClubId> for i16 {
     fn eq(&self, other: &ClubId) -> bool {
         *self == other.0
     }
