+++ fpl_common/src/types/fixture_id.rs
@@ -3,21 +3,21 @@ use std::fmt::Display;
 use std::ops::Deref;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-#[serde(try_from = "i32")]
+#[serde(try_from = "i16")]
 #[derive(sqlx::Type)]
 #[sqlx(transparent)]
-pub struct FixtureId(i32);
+pub struct FixtureId(i16);
 
 #[derive(Debug, thiserror::Error)]
 #[error("FixtureId must be between 1 and 380, got {0}")]
-pub struct FixtureIdError(i32);
+pub struct FixtureIdError(i16);
 
 impl FixtureId {
-    pub const MIN: i32 = 1;
-    pub const MAX: i32 = 380;
-    pub const ALL: std::ops::RangeInclusive<i32> = Self::MIN..=Self::MAX;
+    pub const MIN: i16 = 1;
+    pub const MAX: i16 = 380;
+    pub const ALL: std::ops::RangeInclusive<i16> = Self::MIN..=Self::MAX;
 
-    pub fn new(game_week: i32) -> Result<Self, FixtureIdError> {
+    pub fn new(game_week: i16) -> Result<Self, FixtureIdError> {
         if Self::ALL.contains(&game_week) {
             Ok(Self(game_week))
         } else {
@@ -27,7 +27,7 @@ impl FixtureId {
 }
 
 impl Deref for FixtureId {
-    type Target = i32;
+    type Target = i16;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -40,27 +40,27 @@ impl Display for FixtureId {
     }
 }
 
-impl TryFrom<i32> for FixtureId {
+impl TryFrom<i16> for FixtureId {
     type Error = FixtureIdError;
 
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
+    fn try_from(value: i16) -> Result<Self, Self::Error> {
         Self::new(value)
     }
 }
 
-impl From<FixtureId> for i32 {
+impl From<FixtureId> for i16 {
     fn from(id: FixtureId) -> Self {
         id.0
     }
 }
 
-impl PartialEq<i32> for FixtureId {
-    fn eq(&self, other: &i32) -> bool {
+impl PartialEq<i16> for FixtureId {
+    fn eq(&self, other: &i16) -> bool {
         self.0 == *other
     }
 }
 
-impl PartialEq<FixtureId> for i32 {
+impl PartialEq<FixtureId> for i16 {
     fn eq(&self, other: &FixtureId) -> bool {
         *self == other.0
     }
